requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.14
dj-database-url==2.0.0
dj-rest-auth==4.0.1
dj-rest-auth-social==2.2.4
Django==4.2.1
django-allauth==0.54.0
django-braces==1.15.0
django-cors-headers==3.14.0
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
facebook-sdk==3.1.0
firebase-admin==6.2.0
google==3.0.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
jwcrypto==1.5.0
msgpack==1.0.5
numpy==1.25.2
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
proto-plus==1.22.3
protobuf==4.24.2
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wrapt==1.15.0
